third_party/blink/web_tests/external/wpt/html/semantics/forms/attributes-common-to-form-controls/dirname-only-if-applies-expected.txt

This is a testharness.js-based test.
[FAIL] Submit input element directionality to FormData, if dirname applies.
  assert_equals: Submit ltr for input type=hidden expected (string) "ltr" but got (object) null
[PASS] Submit textarea element directionality to FormData.
[FAIL] Submit input element directionality, if dirname applies.
  assert_equals: Submit ltr for input type=hidden expected (string) "ltr" but got (object) null
[PASS] Submit textarea element directionality.
Harness: the test ran to completion.
